Add handling for scalar returns in torchdynamo refbackend
This commit adds support to the torchdynamo refbackend for e2e tests
that return a scalar value. This issue had not been encountered
previously despite the existence of e2e tests that return scalars
because in those tests torchdynamo is inserting graph breaks that
essentially result in torch-mlir never creating a graph that returns a
scalar. Only certain ops, such as `torch.ops.aten.add` cause
torchdynamo to give a graph with scalar returns to torch-mlir.